modify: index페이지를 main 페이지로 변경(main.tsx 삭제)(#73)
